functions/requirements.txt
firebase-admin
flask
firebase_functions
google.cloud.firestore
